Fixed an issue which crashed the application when the user tried to clear the canvas and there was no sound path.

+++ src/dk/aau/cs/giraf/pictocreator/audiorecorder/AudioHandler.java
@@ -61,6 +61,9 @@ public class AudioHandler {
      * Set the savedFileName to null
      */
     public static void resetSound(){
+        if(savedFileName == null)
+            return;
+
         File tempFile = new File(savedFileName);
         if(tempFile.exists())
             tempFile.delete();
